refactor(cardano-graphql-services): move DataProjector to cardano-graphql-services
