+++ packages/cardano-graphql-services/src/DataProjection/errors/InvalidModuleState.ts
