+++ packages/cardano-graphql-services/src/DataProjection/errors/index.ts
