+++ packages/cardano-graphql-services/src/DataProjection/types.ts
@@ -1,4 +1,4 @@
-import { Asset } from '../Schema/types';
+import { Asset } from '@cardano-graphql/client-ts';
 import { Schema as Ogmios } from '@cardano-ogmios/client';
 import dgraph from 'dgraph-js';
 
@@ -60,42 +60,4 @@ export interface BlockHandler {
   process?: (parameters: ProcessParameters) => Promise<Partial<ProcessingResult>>;
   rollBackward: (ctx: RollBackwardContext) => Promise<Upsert>;
   rollForward: (ctx: RollForwardContext, processingResult: CombinedProcessingResult[]) => Promise<Upsert>;
-}
-
-export interface Signature {
-  signature: string;
-  publicKey: string;
-}
-
-export interface AssetMetadata {
-  decimals: {
-    value: number;
-    anSignatures: Signature[];
-  };
-  description?: {
-    value: string;
-    anSignatures: Signature[];
-  };
-  logo?: {
-    value: string;
-    anSignatures: Signature[];
-  };
-  name?: {
-    value: string;
-    anSignatures: Signature[];
-  };
-  owner?: Signature;
-  preImage?: {
-    value: string;
-    hashFn: string;
-  };
-  subject: string;
-  ticker?: {
-    value: string;
-    anSignatures: Signature[];
-  };
-  url?: {
-    value: string;
-    anSignatures: Signature[];
-  };
 }
