+++ packages/cardano-graphql-services/src/DataProjection/util.ts
