+++ packages/cardano-graphql/src/util/index.ts
@@ -2,4 +2,3 @@ export * from './createProvider';
 export * from './getExactlyOneObject';
 export * from './graphqlPoolParametersToCore';
 export * from './jsonSerializer';
-export * from './logging';
